projects/meeting-assassin/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.13.1

# Authentication & Security
pyjwt==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-oauth2==1.1.1

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# HTTP Client
httpx==0.25.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSocket Support
websockets==12.0

# AI/ML Libraries
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.4

# OpenAI (for AI features) - Enhanced LLM integration
openai==1.48.0
tiktoken==0.5.2  # Token counting for OpenAI models
anthropic==0.34.0  # Anthropic Claude for advanced reasoning

# Speech & Audio Processing
openai-whisper==20231117  # Real-time transcription
pyaudio==0.2.14  # Audio streaming
soundfile==0.12.1
librosa==0.10.1  # Audio analysis
webrtcvad==2.0.10  # Voice activity detection

# Text-to-Speech
elevenlabs==1.3.0  # Voice synthesis
edge-tts==6.1.9  # Microsoft Edge TTS as backup
pyttsx3==2.90  # Offline TTS backup

# WebRTC & Real-time Communication
aiortc==1.6.0  # WebRTC for Python
av==11.0.0  # Audio/Video processing

# Vector Database for Knowledge Base
chromadb==0.4.22
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
faiss-cpu==1.7.4  # Vector similarity search

# Meeting Platform SDKs
zoomus==1.1.7  # Zoom SDK
python-telegram-bot==20.7  # For notifications

# Real-time Processing
asyncio-mqtt==0.16.2
python-socketio[asyncio]==5.11.0
python-engineio==4.8.0

# NLP & Language Processing
spacy==3.7.2
transformers==4.36.2  # Hugging Face transformers
sentence-transformers==2.2.2  # Embeddings

# Computer Vision (for video avatars)
opencv-python==4.9.0.80
mediapipe==0.10.9  # Face detection/tracking

# Data Processing
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0

# Redis for distributed processing
aioredis==2.0.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Date/Time Handling
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
loguru==0.7.2
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
python-dotenv==1.0.0

# Optional: For enhanced features
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2